Removing old implementations and unneeded files
