+++ .vscode/launch.json
@@ -8,7 +8,7 @@
 			"name": "(gdb) Launch",
 			"type": "cppdbg",
 			"request": "launch",
-			"program": "${workspaceFolder}/graph_analysis",
+			"program": "${workspaceFolder}/pmem_benchmark",
 			"args": [],
 			"stopAtEntry": false,
 			"cwd": "${workspaceFolder}",
