+++ include/PMEM/allocator.hpp
@@ -18,6 +18,11 @@ namespace PMEM {
 		}
 
 		T* allocate(std::size_t n) {
+
+			if (n == 0) {
+				return nullptr;
+			}
+
 			if (n > std::numeric_limits<std::size_t>::max() / sizeof(T)) {
 				throw std::bad_alloc();
 			}
