+++ run.sh
@@ -10,7 +10,7 @@ mkdir $out_dir
 export out_dir
 
 # Size of the memory test buffer
-export alloc_size=60000000000
+export alloc_size=2000000000
 
 # Number of vertices to generate for the graph test
 export num_vertices=100000
@@ -24,18 +24,18 @@ export min_value=1
 export max_value=2
 
 # Page Rank Parameters
-export page_rank_iterations=300
+export page_rank_iterations=100
 export page_rank_dampening_factor=0.8
 export num_page_ranks=4
 
 # Number of iterations for each test
-export test_iterations=20
+export test_iterations=5
 
 # OMP environment variables
 export OMP_DISPLAY_ENV=true
 export OMP_PROC_BIND=true
 export OMP_PLACES="{0:36:1}"
-export OMP_NUM_THREADS=36
+export OMP_NUM_THREADS=24
 
 echo Starting pmem_benchmark
 echo Output Directory: $out_dir
