--- src/GraphUtils.cpp
@@ -1,11 +1,0 @@
-#include "GraphUtils.hpp"
-#include <random>
-#include <algorithm>
-#include <iostream>
-#include <fstream>
-#include <sstream>
-#include <functional>
-
-namespace GraphUtils {
-
-}
